pyproject.toml
[project]
name = "memory-rd-strategy-agent"
version = "0.1.0"
description = "HBM4/PIM/CXL R&D strategy analysis agent using multi-agent workflow"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "tavily-python>=0.5.0",
    "sentence-transformers>=3.0.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
eval = [
    "datasets>=2.0.0",
    "pandas>=2.0.0",
]

[project.scripts]
rd-agent = "rd_strategy_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rd_strategy_agent"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
]
